+++ semana20/jest-template/src/performPurchase.ts
@@ -1,0 +1,15 @@
+
+export interface User {
+	name: string
+	balance: number
+}
+
+export const performPurchase = (user: User, value: number): User | undefined => {
+	if(user.balance >= value) {
+		return {
+			...user,
+			balance: user.balance - value		
+		}
+	}
+	return undefined
+}
